bug fix- communication between frontend and backend
