+++ my-app/src/App.css
@@ -73,7 +73,9 @@ nav {
 }
 
 .input-form {
-  left: 2em;
+  left: 1em;
+  border-bottom: 5%;
+  border-radius: 2%;
   padding-left: 10px;
   padding-right: 20em;
   position: relative;
@@ -144,4 +146,6 @@ nav {
 .calculate {
   background-color: #8fe088;
   left: 3em;
+
 }
+
